+++ custom_components/isy994/climate.py
@@ -1,7 +1,6 @@
 """Support for Insteon Thermostats via ISY994 Platform."""
 from __future__ import annotations
 
-
 from pyisy.constants import (
     CMD_CLIMATE_FAN_SETTING,
     CMD_CLIMATE_MODE,
@@ -34,6 +33,7 @@ from homeassistant.const import (
     TEMP_FAHRENHEIT,
 )
 from homeassistant.core import HomeAssistant
+from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from .const import (
     _LOGGER,