+++ custom_components/isy994/fan.py
@@ -5,30 +5,19 @@ import math
 
 from pyisy.constants import ISY_VALUE_UNKNOWN, PROTO_INSTEON
 
-from homeassistant.components.fan import (
-    DOMAIN as FAN,
-    SPEED_HIGH,
-    SPEED_LOW,
-    SPEED_MEDIUM,
-    SPEED_OFF,
-    SUPPORT_SET_SPEED,
-    FanEntity,
-)
+from homeassistant.components.fan import DOMAIN as FAN, SUPPORT_SET_SPEED, FanEntity
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
-
-from .const import _LOGGER, DOMAIN as ISY994_DOMAIN, ISY994_NODES, ISY994_PROGRAMS
-from .entity import ISYNodeEntity, ISYProgramEntity
-from .helpers import migrate_old_unique_ids
-from .percentage import (
+from homeassistant.helpers.entity_platform import AddEntitiesCallback
+from homeassistant.util.percentage import (
     int_states_in_range,
-    ordered_list_item_to_percentage,
-    percentage_to_ordered_list_item,
     percentage_to_ranged_value,
     ranged_value_to_percentage,
 )
 
-ON_SPEEDS = [SPEED_LOW, SPEED_MEDIUM, SPEED_HIGH]
+from .const import _LOGGER, DOMAIN as ISY994_DOMAIN, ISY994_NODES, ISY994_PROGRAMS
+from .entity import ISYNodeEntity, ISYProgramEntity
+from .helpers import migrate_old_unique_ids
 
 SPEED_RANGE = (1, 255)  # off is not included
 
@@ -56,16 +45,6 @@ class ISYFanEntity(ISYNodeEntity, FanEntity):
     """Representation of an ISY994 fan device."""
 
     @property
-    def speed(self) -> str:
-        """Return the current speed."""
-        pecentage = self.percentage
-        if pecentage is None:
-            return None
-        if pecentage == 0:
-            return SPEED_OFF
-        return percentage_to_ordered_list_item(ON_SPEEDS, self.percentage)
-
-    @property
     def percentage(self) -> int | None:
         """Return the current speed percentage."""
         if self._node.status == ISY_VALUE_UNKNOWN:
@@ -75,6 +54,8 @@ class ISYFanEntity(ISYNodeEntity, FanEntity):
     @property
     def speed_count(self) -> int:
         """Return the number of speeds the fan supports."""
+        if self._node.protocol == PROTO_INSTEON:
+            return 3
         return int_states_in_range(SPEED_RANGE)
 
     @property
@@ -94,12 +75,6 @@ class ISYFanEntity(ISYNodeEntity, FanEntity):
 
         await self._node.turn_on(val=isy_speed)
 
-    async def async_set_speed(self, speed: str) -> None:
-        """Send the set speed command to the ISY994 fan device."""
-        await self.async_set_percentage(
-            ordered_list_item_to_percentage(ON_SPEEDS, speed)
-        )
-
     async def async_turn_on(
         self,
         speed: str = None,
@@ -108,19 +83,11 @@ class ISYFanEntity(ISYNodeEntity, FanEntity):
         **kwargs,
     ) -> None:
         """Send the turn on command to the ISY994 fan device."""
-        if percentage is not None:
-            await self.async_set_percentage(percentage)
-        else:
-            await self.async_set_speed(speed)
+        await self.async_set_percentage(percentage)
 
     async def async_turn_off(self, **kwargs) -> None:
         """Send the turn off command to the ISY994 fan device."""
         await self._node.turn_off()
-
-    @property
-    def speed_list(self) -> list:
-        """Get the list of available speeds."""
-        return [SPEED_OFF, *ON_SPEEDS]
 
     @property
     def supported_features(self) -> int:
@@ -132,16 +99,6 @@ class ISYFanProgramEntity(ISYProgramEntity, FanEntity):
     """Representation of an ISY994 fan program."""
 
     @property
-    def speed(self) -> str:
-        """Return the current speed."""
-        pecentage = self.percentage
-        if pecentage is None:
-            return None
-        if pecentage == 0:
-            return SPEED_OFF
-        return percentage_to_ordered_list_item(ON_SPEEDS, self.percentage)
-
-    @property
     def percentage(self) -> int | None:
         """Return the current speed percentage."""
         if self._node.status == ISY_VALUE_UNKNOWN:
@@ -151,8 +108,6 @@ class ISYFanProgramEntity(ISYProgramEntity, FanEntity):
     @property
     def speed_count(self) -> int:
         """Return the number of speeds the fan supports."""
-        if self._node.protocol == PROTO_INSTEON:
-            return 3
         return int_states_in_range(SPEED_RANGE)
 
     @property